# go-architecture/hexagonal-architecture/go-hexagonal-architecture-2/docs/swagger.yaml
basePath: /v1
definitions:
...
  handler.authResponse:
    properties:
      token:
        example: v2.local.Gdh5kiOTyyaQ3_bNykYDeYHO21Jg2...
        type: string
    type: object
  handler.categoryResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Foods
        type: string
    type: object
  handler.createCategoryRequest:
    properties:
      name:
        example: Foods
        type: string
    required:
    - name
    type: object
...
  handler.errorResponse:
    properties:
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  handler.loginRequest:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  handler.meta:
    properties:
      limit:
        example: 10
        type: integer
      skip:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  handler.response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.updateCategoryRequest:
    properties:
      name:
        example: Beverages
        type: string
    required:
    - name
    type: object
...
host: gopos.bagashiz.me
info:
  contact:
    name: test
    url: test
  description: This is a simple RESTful Point of Sale (POS) Service API written in
    Go using Gin web framework, PostgreSQL database, and Redis cache.
  license:
    name: MIT
    url: test
  title: Go POS (Point of Sale) API
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: List categories with pagination
      parameters:
      - description: Skip
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Categories displayed
          schema:
            $ref: '#/definitions/handler.meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: create a new category with name
      parameters:
      - description: Create category request
        in: body
        name: createCategoryRequest
        required: true
        schema:
          $ref: '#/definitions/handler.createCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category created
          schema:
            $ref: '#/definitions/handler.categoryResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted
          schema:
            $ref: '#/definitions/handler.response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: get a category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category retrieved
          schema:
            $ref: '#/definitions/handler.categoryResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: update a category's name by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update category request
        in: body
        name: updateCategoryRequest
        required: true
        schema:
          $ref: '#/definitions/handler.updateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated
          schema:
            $ref: '#/definitions/handler.categoryResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Logs in a registered user and returns an access token if the credentials
        are valid.
      parameters:
      - description: Login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Succesfully logged in
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Login and get an access token
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"